fix: 홈 통화 데이터 처리 안정화

+++ core/network/src/main/java/kr/co/call/network/dto/home/CallTranscriptDto.kt
@@ -1,7 +1,7 @@
 package kr.co.call.network.dto.home
 
 data class CallTranscriptDto(
-    val callId: Int,
+    val callId: Long,
     val lines: List<CallScriptMessage>,
 )
